requirements-dev.txt
-r requirements.txt
alembic==1.16.2
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.1
pytest==8.4.1
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1
